code refactoring fix
