+++ src/main/java/com/example/backend/security/CorsConfig.java
@@ -1,0 +1,23 @@
+package com.example.backend.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
+
+import java.util.Arrays;
+
+@Component
+public class CorsConfig {
+    @Bean
+    public CorsFilter corsFilter() {
+        UrlBasedCorsConfigurationSource urlBasedCorsConfigurationSource = new UrlBasedCorsConfigurationSource();
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(Arrays.asList("http://localhost:3000","https://marketplace-frontend-mu.vercel.app/"));
+        config.addAllowedMethod("*");
+        config.addAllowedHeader("*");
+        urlBasedCorsConfigurationSource.registerCorsConfiguration("/**", config);
+        return new CorsFilter(urlBasedCorsConfigurationSource);
+    }
+}
