+++ src/main/java/com/example/backend/web/File/FileUpload.java
@@ -2,10 +2,9 @@ package com.example.backend.web.File;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.Map;
 
 public interface FileUpload {
-    Map uploadFile(MultipartFile multipartFile) throws IOException;
+    Map uploadFile(MultipartFile multipartFile);
     Map deleteById(String id);
 }
